requirements.txt
hikari-lightbulb==0.0.59
hikari==2.0.0.dev100
python-dotenv
pillow-simd; platform_system == "Linux"
pillow; platform_system == "Windows"
git+https://github.com/Supersebi3/pilutils/